+++ src/main/java/ortus/boxlang/modules/orm/interceptors/TransactionManager.java
@@ -65,11 +65,14 @@ public class TransactionManager extends BaseInterceptor {
 			    "No ORM application found during transaction request.  Either the ORM service is not properly configured or the application has not yet started." );
 			return;
 		}
-		ORMContext	ormContext	= ORMContext.getForContext( context.getParentOfType( IJDBCCapableContext.class ) );
-		ORMConfig	config		= ormContext.getConfig();
+		IJDBCCapableContext	jdbcContext	= context.getParentOfType( IJDBCCapableContext.class );
+		ORMContext			ormContext	= ORMContext.getForContext( jdbcContext );
+		ORMConfig			config		= ormContext.getConfig();
 
 		ormApp.getDatasources().forEach( ( datasource ) -> {
 			Session ormSession = ormContext.getSession( datasource );
+			// Ensure any pending operations are flushed before starting the transaction
+			ormSession.flush();
 			// We should never hit this conditional as long as BoxLang does not support nested transactions
 			if ( ormSession.isJoinedToTransaction() ) {
 				if ( logger.isDebugEnabled() ) {